third_party/blink/web_tests/external/wpt/fetch/data-urls/processing.any.serviceworker-expected.txt

This is a testharness.js-based test.
Found 71 tests; 49 PASS, 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Setup.
PASS "data://test/,X"
FAIL "data://test:test/,X" assert_unreached: Should have rejected: undefined Reached unreachable code
PASS "data:,X"
PASS "data:"
PASS "data:text/html"
PASS "data:text/html    ;charset=x   "
PASS "data:,"
PASS "data:,X#X"
PASS "data:,%FF"
PASS "data:text/plain,X"
PASS "data:text/plain ,X"
PASS "data:text/plain%20,X"
PASS "data:text/plain\f,X"
PASS "data:text/plain%0C,X"
PASS "data:text/plain;,X"
FAIL "data:;x=x;charset=x,X" assert_equals: expected "text/plain;x=x;charset=x" but got "text/plain;charset=x"
FAIL "data:;x=x,X" assert_equals: expected "text/plain;x=x" but got "text/plain;charset=US-ASCII"
PASS "data:text/plain;charset=windows-1252,%C2%B1"
PASS "data:text/plain;Charset=UTF-8,%C2%B1"
PASS "data:image/gif,%C2%B1"
PASS "data:IMAGE/gif,%C2%B1"
FAIL "data:IMAGE/gif;hi=x,%C2%B1" assert_equals: expected "image/gif;hi=x" but got "image/gif"
PASS "data:IMAGE/gif;CHARSET=x,%C2%B1"
PASS "data: ,%FF"
PASS "data:%20,%FF"
PASS "data:\f,%FF"
PASS "data:%1F,%FF"
PASS "data:\0,%FF"
PASS "data:%00,%FF"
PASS "data:text/html  ,X"
PASS "data:text / html,X"
PASS "data:†,X"
PASS "data:†/†,X"
PASS "data:X,X"
FAIL "data:image/png,X X" assert_array_equals: lengths differ, expected array [88, 32, 88] length 3, got object "88,88" length 2
FAIL "data:application/javascript,X X" assert_array_equals: lengths differ, expected array [88, 32, 88] length 3, got object "88,88" length 2
PASS "data:application/xml,X X"
PASS "data:text/javascript,X X"
PASS "data:text/plain,X X"
FAIL "data:unknown/unknown,X X" assert_array_equals: lengths differ, expected array [88, 32, 88] length 3, got object "88,88" length 2
FAIL "data:text/plain;a=\",\",X" assert_equals: expected "text/plain;a=\"\"" but got "text/plain"
FAIL "data:text/plain;a=%2C,X" assert_equals: expected "text/plain;a=%2C" but got "text/plain"
FAIL "data:;base64;base64,WA" assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
PASS "data:x/x;base64;base64,WA"
FAIL "data:x/x;base64;charset=x,WA" assert_array_equals: lengths differ, expected array [87, 65] length 2, got object "88" length 1
PASS "data:x/x;base64;charset=x;base64,WA"
FAIL "data:x/x;base64;base64x,WA" assert_array_equals: lengths differ, expected array [87, 65] length 2, got object "88" length 1
PASS "data:;base64,W%20A"
PASS "data:;base64,W%0CA"
PASS "data:x;base64x,WA"
FAIL "data:x;base64;x,WA" assert_array_equals: lengths differ, expected array [87, 65] length 2, got object "88" length 1
PASS "data:x;base64=x,WA"
PASS "data:; base64,WA"
PASS "data:;  base64,WA"
PASS "data:  ;charset=x   ;  base64,WA"
FAIL "data:;base64;,WA" assert_array_equals: lengths differ, expected array [87, 65] length 2, got object "88" length 1
PASS "data:;base64 ,WA"
PASS "data:;base64   ,WA"
FAIL "data:;base 64,WA" assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
PASS "data:;BASe64,WA"
FAIL "data:;%62ase64,WA" assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
PASS "data:%3Bbase64,WA"
PASS "data:;charset=x,X"
PASS "data:; charset=x,X"
FAIL "data:;charset =x,X" assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
FAIL "data:;charset= x,X" promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
FAIL "data:;charset=,X" promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
FAIL "data:;charset,X" assert_equals: expected "text/plain" but got "text/plain;charset=US-ASCII"
FAIL "data:;charset=\"x\",X" promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
FAIL "data:;CHARSET=\"X\",X" promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
Harness: the test ran to completion.
